// services/te-base-api/src/utils/enums.ts
export enum SortOrder {
  asc = "asc",
  desc = "desc"
}

export enum JobStatus {
  HIRED = "hired",
  PENDING = "pending",
  COMPLETED = "completed"
}

export enum JobRequestStatus {
  ACTIVE = "active",
  INACTIVE = "inactive",
}

export enum UserRoles {
  EMPLOYEE = "Employee",
  EMPLOYER = "Employer",
  SUPPER_ADMIN = "Super Admin"
}

export enum SearchType {
  COUNT = 'count',
  DATA = 'data',
}

export enum ReviewSortTypes {
  MOST_RELEVANT = "most relevant",
  MOST_RECENT = "most recent",
  MOST_COMPLIMENTARY = "most complimentary",
  MOST_CRITICAL = "most critical"
}

export enum BidStatus {
  PENDING = "pending",
  ACCEPTED = "accepted",
  REJECTED = "rejected"
}

export enum BooleanTypes {
  TRUE = "true",
  FALSE = "false"
}